requirements.txt
selenium==4.36.0
pydantic==2.11.10
python-dotenv==1.1.1
sqlalchemy
psycopg2
alembic==1.16.5
